requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
passlib
bcrypt
argon2_cffi
python-jose
python-multipart